requirement.txt
certifi==2018.11.29
chainer==5.1.0
chainercv==0.11.0
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29.1
dask==1.0.0
decorator==4.3.0
fcn==6.4.8
filelock==3.0.10
gdown==3.6.0
idna==2.7
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
pandas==0.23.4
Pillow==5.3.0
protobuf==3.6.1
pycocotools==2.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
requests==2.20.1
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
toolz==0.9.0
torch==0.4.1
torchfcn==1.9.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
urllib3==1.24.1
visdom==0.1.8.5
websocket-client==0.54.0
